integrating new perception changes
