+++ GEMstack/knowledge/defaults/computation_graph.yaml
@@ -18,11 +18,14 @@ components:
   - agent_detection:
       inputs: vehicle
       outputs: agents
+  - obstacle_detection:
+      inputs: vehicle
+      outputs: obstacles
   - lane_detection:
       inputs: [vehicle, roadgraph]
       outputs: vehicle_lane
   - parking_detection:
-      inputs: agents
+      inputs: obstacles
       outputs: [goal, obstacles]
   - sign_detection: 
       inputs: [vehicle, roadgraph]
