+++ GEMstack/state/__init__.py
@@ -12,6 +12,7 @@ __all__ = ['PhysicalObject','ObjectPose','ObjectFrameEnum',
            'Sign',
            'AgentState','AgentEnum','AgentActivityEnum',
            'SceneState',
+           'ObstacleState','ObstacleMaterialEnum','ObstacleStateEnum',
            'VehicleIntent','VehicleIntentEnum',
            'AgentIntent',
            'EntityRelationEnum','EntityRelation','EntityRelationGraph',
@@ -24,6 +25,7 @@ from .physical_object import PhysicalObject, ObjectPose, ObjectFrameEnum
 from .trajectory import Path,Trajectory
 from .vehicle import VehicleState,VehicleGearEnum
 from .roadgraph import Roadgraph, RoadgraphLane, RoadgraphCurve, RoadgraphRegion, RoadgraphCurveEnum, RoadgraphLaneEnum, RoadgraphRegionEnum, RoadgraphSurfaceEnum, RoadgraphConnectionEnum
+from .obstacle import Obstacle, ObstacleMaterialEnum, ObstacleState, ObstacleStateEnum
 from .obstacle import Obstacle, ObstacleMaterialEnum
 from .sign import Sign, SignEnum, SignalLightEnum, SignState
 from .roadmap import Roadmap
